+++ back/src/main/java/com/siban/back/board/controller/PostController.java
@@ -1,0 +1,5 @@
+package com.siban.back.board.controller;
+
+public class PostController {
+
+}
