+++ back/src/main/java/com/siban/back/board/controller/ReplyController.java
@@ -1,0 +1,5 @@
+package com.siban.back.board.controller;
+
+public class ReplyController {
+
+}
