+++ back/src/main/java/com/siban/back/board/domain/Post.java
@@ -1,4 +1,4 @@
-package com.siban.board.domain;
+package com.siban.back.board.domain;
 
 import javax.persistence.Entity;
 import javax.persistence.Table;
