+++ back/src/main/java/com/siban/back/board/domain/Reply.java
@@ -1,0 +1,5 @@
+package com.siban.back.board.domain;
+
+public class Reply {
+
+}
