+++ back/src/main/java/com/siban/back/board/repository/ReplyRepository.java
@@ -1,0 +1,5 @@
+package com.siban.back.board.repository;
+
+public interface ReplyRepository {
+
+}
