+++ back/src/main/java/com/siban/back/board/service/PostService.java
@@ -1,0 +1,7 @@
+package com.siban.back.board.service;
+
+import com.siban.back.board.repository.PostRepository;
+
+public class PostService implements PostRepository {
+
+}
