+++ back/src/main/java/com/siban/back/board/service/ReplyService.java
@@ -1,0 +1,7 @@
+package com.siban.back.board.service;
+
+import com.siban.back.board.repository.ReplyRepository;
+
+public class ReplyService implements ReplyRepository {
+
+}
