--- back/src/main/java/com/siban/board/controller/PostController.java
@@ -1,5 +1,0 @@
-package com.siban.board.controller;
-
-public class PostController {
-
-}
