--- back/src/main/java/com/siban/board/domain/Reply.java
@@ -1,5 +1,0 @@
-package com.siban.board.domain;
-
-public class Reply {
-
-}
