--- back/src/main/java/com/siban/board/repository/PostRepository.java
@@ -1,5 +1,0 @@
-package com.siban.board.repository;
-
-public interface PostRepository {
-
-}
