--- back/src/main/java/com/siban/board/repository/ReplyRepository.java
@@ -1,5 +1,0 @@
-package com.siban.board.repository;
-
-public interface ReplyRepository {
-
-}
