--- back/src/main/java/com/siban/board/service/PostService.java
@@ -1,7 +1,0 @@
-package com.siban.board.service;
-
-import com.siban.board.repository.PostRepository;
-
-public class PostService implements PostRepository {
-
-}
