--- back/src/main/java/com/siban/board/service/ReplyService.java
@@ -1,7 +1,0 @@
-package com.siban.board.service;
-
-import com.siban.board.repository.ReplyRepository;
-
-public class ReplyService implements ReplyRepository {
-
-}
